Revision of AppFrame layout
Big update for the UI of AppFrame.
pretty much flipped the order on it's head -
moved all the of old JTextFields to a nested class that is now called from AppFrame using just the employee ID.
All text fields are prepopulated from a new getEmployeeData method in the Employee class.
This means you don't  have to type the employees complete information each time, just edit the existing.

Also overloaded the printTableData method in Employee class to take an argument to print a single row of information if argument is provided.